+++ src/index.ts
@@ -9,7 +9,7 @@ export {
 export { Field, Bool } from './lib/core.js';
 export type { VerificationKey, Keypair } from './snarky.js';
 export * from './snarky/addons.js';
-export { Poseidon } from './lib/hash.js';
+export { Poseidon, TokenSymbol } from './lib/hash.js';
 export * from './lib/signature.js';
 export {
   Circuit,
@@ -32,6 +32,7 @@ export {
   DeployArgs,
   signFeePayer,
   declareMethods,
+  Account,
 } from './lib/zkapp.js';
 export { state, State, declareState } from './lib/state.js';
 export { Proof, SelfProof, ZkProgram, verify } from './lib/proof_system.js';
@@ -56,7 +57,10 @@ export { Character, CircuitString } from './lib/string.js';
 
 // experimental APIs
 import { Reducer, Callback, accountUpdateFromCallback } from './lib/zkapp.js';
-import { createChildAccountUpdate } from './lib/account_update.js';
+import {
+  createChildAccountUpdate,
+  makeChildAccountUpdate,
+} from './lib/account_update.js';
 import {
   memoizeWitness,
   AsFieldsAndAux as AsFieldsAndAux_,
@@ -69,12 +73,13 @@ const Experimental_ = {
   Callback,
   accountUpdateFromCallback,
   createChildAccountUpdate,
+  makeChildAccountUpdate,
   memoizeWitness,
   MerkleTree,
   MerkleWitness,
 };
 
-type Callback_ = Callback;
+type Callback_<Result> = Callback<Result>;
 
 /**
  * This module exposes APIs that are unstable, in the sense that the API surface is expected to change.
@@ -83,11 +88,13 @@ type Callback_ = Callback;
 namespace Experimental {
   export let Reducer = Experimental_.Reducer;
   export let createChildAccountUpdate = Experimental_.createChildAccountUpdate;
+  export let makeChildAccountUpdate = Experimental_.makeChildAccountUpdate;
   export let memoizeWitness = Experimental_.memoizeWitness;
   export let MerkleTree = Experimental_.MerkleTree;
   export let MerkleWitness = Experimental_.MerkleWitness;
-  export let accountUpdateFromCallback = Experimental_.accountUpdateFromCallback;
+  export let accountUpdateFromCallback =
+    Experimental_.accountUpdateFromCallback;
   export type AsFieldsAndAux<T, TJson> = AsFieldsAndAux_<T, TJson>;
   export let Callback = Experimental_.Callback;
-  export type Callback = Callback_;
+  export type Callback<Result> = Callback_<Result>;
 }