+++ src/lib/precondition.ts
@@ -8,10 +8,11 @@ import {
   Preconditions,
 } from './account_update.js';
 import { UInt32, UInt64 } from './int.js';
-import { inAnalyze, inCompile, inProver } from './proof_system.js';
+import { inCompileMode, inProver } from './proof_system.js';
 import { Layout } from '../snarky/parties-helpers.js';
 import { jsLayout } from '../snarky/types.js';
 import { emptyReceiptChainHash } from './hash.js';
+import { PublicKey } from './signature.js';
 
 export {
   preconditions,
@@ -21,6 +22,7 @@ export {
   cleanPreconditionsCache,
   AccountValue,
   NetworkValue,
+  getAccountPreconditions,
 };
 
 function preconditions(accountUpdate: AccountUpdate, isSelf: boolean) {
@@ -170,7 +172,7 @@ function getVariable<K extends LongKey, U extends FlatPreconditionValue[K]>(
   fieldType: AsFieldElements<U>
 ): U {
   // in compile, just return an empty variable
-  if (inCompile() || inAnalyze()) {
+  if (inCompileMode()) {
     return Circuit.witness(fieldType, (): U => {
       // TODO this error is never thrown. instead, reading the value with e.g. `toString` ends up
       // calling snarky's eval_as_prover, which throws "Can't evaluate prover code outside an as_prover block"
@@ -186,7 +188,7 @@ To write a correct circuit, you must avoid any dependency on the concrete value 
   let key = rest.join('.');
   let value: U;
   if (accountOrNetwork === 'account') {
-    let account = getAccountPreconditions(accountUpdate);
+    let account = getAccountPreconditions(accountUpdate.body);
     value = account[key as keyof AccountValue] as U;
   } else if (accountOrNetwork === 'network') {
     let networkState = Mina.getNetworkState();
@@ -203,8 +205,11 @@ To write a correct circuit, you must avoid any dependency on the concrete value 
   }
 }
 
-function getAccountPreconditions(accountUpdate: AccountUpdate): AccountValue {
-  let { publicKey, tokenId } = accountUpdate.body;
+function getAccountPreconditions(body: {
+  publicKey: PublicKey;
+  tokenId?: Field;
+}): AccountValue {
+  let { publicKey, tokenId } = body;
   let hasAccount = Mina.hasAccount(publicKey, tokenId);
   if (!hasAccount) {
     return {